feat: Add seat swapping functionality to admin panel
This commit introduces a new feature that allows administrators to swap seats in the admin panel.  It includes new components and actions for managing seat swaps.
